icyci: attempt to gracefully shutdown on SIGTERM
I have a unit test for this which intermittently times out, so there's
still some work to be done here. There appears to be a race condition in
awaitCommand() cmd.Wait() handling between goroutines.

This is still a step in the right direction though, so I'll merge it for
now.

Signed-off-by: David Disseldorp <ddiss@suse.de>