test: reuse eventLoop pollInterval for waitNotes
This should allow us to reduce some of the sleep time during testing.
Ideally in future we will replace waitNotes() polling with a git
post-receive hook which signals the test process once notes arrive.

TestSeparateSrcRslt is a special case, as the (long) eventLoop
pollInterval isn't used by the test: we're only interested in waiting
for the notes from the first (on-start) test-script run.

Signed-off-by: David Disseldorp <ddiss@suse.de>